Add cubari types and clean up
Modify the RemoteStorage initializer so its easier to read. Also add
types for the Cubari module since that wasn't present by default.

The remotestoragejs module can't be augmented, so an extended version
of the RemoteStorage class had to be created. An issue has been made
in the package repository regarding this.

Signed-off-by: kingbri <bdashore3@gmail.com>